fixed compiler warnings for fecTagUntagBooster.c;

+++ cpuBoosters/fecEncodeBooster/fecTagUntagBooster.c
@@ -12,11 +12,11 @@ void my_packet_handler(
 	u_char *new_packet = NULL;
 	int tagged_size = wharf_tag_frame(packet, header->len, &new_packet);
 	int new_size = wharf_strip_frame(new_packet, tagged_size);
-	if (new_size != header->len) {
+	if ((unsigned)new_size != header->len) {
 		fprintf(stderr, "Frame size changed\n");
 		exit(1);
 	}
-	for (int i = 0; i < header->len; i++) {
+	for (int i = 0; (unsigned)i < header->len; i++) {
 		if (new_packet[i] != packet[i]) {
 			fprintf(stderr, "Frame changed\n");
 			exit(1);
